opentcs-common/src/test/java/org/opentcs/util/gui/StringTableCellRendererTest.java
// SPDX-FileCopyrightText: The openTCS Authors
// SPDX-License-Identifier: MIT
package org.opentcs.util.gui;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.instanceOf;
import static org.hamcrest.Matchers.is;

import java.awt.Component;
import javax.swing.JLabel;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.opentcs.data.TCSObjectReference;
import org.opentcs.data.model.Vehicle;

/**
 */
class StringTableCellRendererTest {

  private StringTableCellRenderer<TCSObjectReference<?>> renderer;

  @BeforeEach
  void setUp() {
    renderer = new StringTableCellRenderer<>(x -> x == null ? "" : x.getName());
  }

  @Test
  void returnsNullForNullValue() {
    Component result = renderer.getTableCellRendererComponent(null, null, false, false, 0, 0);
    assertThat(result, is(instanceOf(JLabel.class)));
    JLabel labelResult = (JLabel) result;
    assertThat(labelResult.getText(), is(equalTo("")));
  }

  @Test
  void returnsLabelWithNameAsText() {
    Vehicle vehicle = new Vehicle("VehicleName");
    TCSObjectReference<Vehicle> vehicleReference = vehicle.getReference();
    Component result
        = renderer.getTableCellRendererComponent(null, vehicleReference, false, false, 0, 0);
    assertThat(result, is(instanceOf(JLabel.class)));
    JLabel labelResult = (JLabel) result;
    assertThat(labelResult.getText(), is(equalTo("VehicleName")));
  }

}
